so many things
base projectile, dash (for now), groundwork for black hole attack

+++ Source/ProjectLightman/Private/PLMAttributeComponent.cpp
@@ -31,7 +31,7 @@ void UPLMAttributeComponent::TickComponent(float DeltaTime, ELevelTick TickType,
 	// ...
 }
 
-bool UPLMAttributeComponent::ApplyHealthChange(float Delta)
+bool UPLMAttributeComponent::ApplyHealthChange(int Delta)
 {
 	Health += Delta;
